Trata campo número

+++ app/Services/Educacenso/Version2019/Registro00Import.php
@@ -208,7 +208,7 @@ class Registro00Import implements RegistroImportInterface
             'tipo' => '1',
             'cep' => $this->model->cep,
             'idlog' => $street->getKey(),
-            'numero' => is_numeric($this->model->numero) ? $this->model->numero : null,
+            'numero' => (int) (is_numeric($this->model->numero) ? $this->model->numero : null),
             'complemento' => $this->model->complemento,
             'idbai' => $neighborhood->getKey(),
             'origem_gravacao' => 'M',
